ENH: improvements to the event broker, and conversion of many internal classes to use the event broker.  See https://www.slicer.org/wiki/Slicer3:EventBroker
svn-url: http://viewvc.slicer.org/viewvc.cgi/Slicer4?view=revision&revision=6536
git-svn-id: http://svn.slicer.org/Slicer4/trunk@6536 3bd1e089-480b-0410-8dfb-8563597acbee

+++ Libs/MRML/vtkMRMLColorNode.cxx
@@ -188,7 +188,7 @@ void vtkMRMLColorNode::PrintSelf(ostream& os, vtkIndent indent)
   if (this->Names.size() > 0)
     {
     os << indent << "Color Names:\n";
-    for (unsigned int i = 0; i < this->Names.size(); i++)
+    for (unsigned int i = 0; (int)i < this->Names.size(); i++)
       {
       os << indent << indent << i << " " << this->GetColorName(i) << endl;
       if ( i > 10 )
